+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ## Neste versjon
 
 - ✨ **Pages**. Lagt til mulighet til å søke innad i pages.
+- ✨ **Profil**. Pagination på arrangementer og badges, samt lagt til dine grupper.
 - 🎨 **Nyheter**. Nyheter viser nå hvilket år de ble publisert
 
 ## Versjon 1.0.9 (05.05.2021)